Remove the "XyceAsLibrary" tests
There were two binaries that could be built in the XyceAsLibrary directory. The
"testHarness1685" test was part of a regression test that has been removed. The
"bug1574" code does not compile as-is, and is not tested, so it does not serve
any purpose. It can always be resurrected via history, if needed.

Xyce/backlog/bugs#3