Add BSIM 4.7.0 to Third Party Licenses file
When we added the BSIM 4.7.0, we should have noted its license is the
same as the other early BSIM devices.

I am getting to this, because when time comes to add BSIM 4.8.2, that
license is different and this file will need updating again.  I was
just hunting down where to add the updated license when I noticed that
we forgot the 4.7.0 addition.

Xyce/backlog/project-backlog#420