Add `Xyce_TEST_BINARIES` CMake configure flag
This adds a CMake configuration variable, Xyce_TEST_BINARIES. It controls
whether the the test binaries for Xyce_Regression are built, and is ON by
default. It revises the previous BUILD_TESTING commit, in that it changes the
variable that controls when the binaries are built. However, the "unit" tests
associated with these binaries are still controlled by BUILD_TESTING.

Xyce/backlog/bugs#3