Add Auger and SRH options to 2D TCAD device params
I recently fixed a scaling mistake in the PDE (TCAD) devices for Auger
recombination.  It broke several tests.

One of the tests that it broke was the PDE_2D_BJT/invertbjt.cir test.
This test consists of 2 BJT TCAD devices (an NPN and a PNP), and as it
is a somewhat slow-running test it was only run weekly.  My fix for that
test was to add "auger=false" to the parameters of both devices.
However, this initially faied b/c I had never added the "AUGER"
parameter to the 2D device.  I had only added it to the 1D devices.

As of this commit, AUGER=true/false is now supported in the 2D devices.

Also added in this commit is SRH=true/false as well as the user-defined
scaling parameters C0 and t0.  Initially after my fix the 2D device was
having convergence problems, and sometimes adjusting the concentration
scaling (C0) can help.  So, I added that parameter as part of
this work.

The convergence problem turned out to be unrelated to scaling, and
didn't improve with different C0 values.  It turned out I'd made a code
mistake, since fixed.  But, having C0 available is very useful (I've had
it in the 1D devices for a long time), so I am including it in this
commit.

Xyce/backlog/bugs#9
Xyce/backlog/project-backlog#448