Make XyceCInterface install components optional
The XyceCInterface is a optional build component but the
install command did not assume it was optional.  This commit
makes the install optional as well.  If XyceCInterface is
build then it will be installed.  If it is not built
then there will not be an error that it's missing.

ihttps://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/3