Add BSIM4 version 4.8.2
With this commit, Xyce can select BSIM 4.8.2 code based on the setting
of the model card's VERSION parameter.

If the user requests VERSION<4.61, the user is warned that they have
requested an older version than any supported and the code
uses 4.61.

If the user requests (VERSION>=4.61 && VERSION<4.70), then the user
gets 4.61.  If they did not request 4.61, they are warned that they're
getting 4.61.

If the user requests (VERSION>=4.70 VERSION<4.80) then they get 4.70.
If they didn't request 4.70, they are warned that's what they're
getting.

If the user requests (VERSION>=4.80) they get 4.82.  If they didn't
request 4.82, they are warned they're getting 4.82.
If they request VERSION>4.82, they are warned that they have asked for
too new a version, and they're getting the older 4.82 instead.

Note that the 4.8.2 code has conditionals that make it do slightly
different things if version is <= 4.80.  This means that our choice to
use 4.8.2 code when the user requests 4.8.0 is the right choice,
because the code knows the difference.

This code has not yet been fully tested.  I am slowly picking through
the full set of BSIM 4.8.1 test cases (the latest provided by the BSIM
group) to assure that it passes all of them.  I will also run this
through the CMC test suite before declaring victory.

Xyce/backlog/project-backlog#352