Fixes testBlockLinearSystems.C to be compilable
This test has never been correctly integrated in the testing infrastructure,
so it is no surprise that it wouldn't compile.  The test essentially serves
as a unit test for the block linear algebra infrastructure.  It reads in a
Matrix Market file and creates block matrices and vectors with it, then solves
the linear system using Amesos.

There was some additional block matrix/vector constructors being used in this
test that are no longer necessary, so they have been removed.  Also the input
Matrix Market files need to be moved to the Xyce_Regression directory for
testing, so they are being removed from here.

Xyce/backlog/bugs#14