Update a i/vsrc error message to be meaningful
Xyce/backlog/project-backlog#455
"Xyce gives a bad error message for malformed voltage source"

I was recently working with a benchmark circuit.  It happened to have a
grammatical mistake in a Vsrc device.  Xyce emitted an error and exited,
but the error was completely uninformative.

The voltage source that caused the problem looked like this:

V2 1 0 dc 1.0 dc 0.0

The problem is that it specified two DC values.  In some codes this is
fine; it will just use one of them and ignore the other.  But Xyce isn't
able to parse it.  Possibly we should fix that.

But in the meantime, the error message needs to be improved.  Prior to
this commit, the error message looked like this:

***** Reading and parsing netlist...
No such source function dc
*** Xyce Abort ***
No such source function dc

No information about which device was causing the error, or the line
number, or the netlist file name.

With this commit, this has been updated to:

***** Reading and parsing netlist...
Netlist error in file small.cir at or near line 4
 No such source function dc in V2
Netlist error: Simulation aborted due to error.  There are 0 MSG_FATAL errors
 and 1 MSG_ERROR errors

This fix is an improvement, but isn't perfect, in that most users
would consider "dc" to be a valid source function.  But at the point
that this error is emitted, it is exclusively looking for transient
sources.

The best most comprehensive fix would be for Xyce to coherently handle
multiple DC values, and just use the last one.