Update voltage node text
This commit is in support of:

 Xyce/backlog/project-backlog#451
"Allow internal subcircuit nodes to be accessed from netlist top level
(not via interface nodes)"

THe text claimed that only .PRINT lines and also Bsrcs could access
internal subcircuit nodes (using the fully resolved syntax).  However,
this is out of date; conventional nodal access on device instances
lines also works.

One of our users had an important use case for this capability, but
assumed it wouldn't work b/c of the text in the guides.  This commit
corrects that text.