Add probe for suitesparseconfig
Newer AMD libraries depend on suitesparseconfig.

If only static libraries are available, this means that you MUST probe
for suitesparseconfig before probing for amd.  On systems where there
are shared libraries only, that is not necessary because the
dependency is already coded in the library itself.

This has never been a problem before on any system where we've built
Xyce, but it could be the source of a problem that is coming up now
for a user.  Maybe not.