Adds an optional argument to the SSFT::getLine method
The SpiceSeparateFieldTool::getLine method can now take an optional
argument, which is a vector of strings.  This vector of strings, if
provided, will be used by the tokenizer to select which lines are
fully tokenized by the getLine method.  This method can be used by
the CircuitBlock::parsePreprocess method and DistToolBase::getLine
method to more selectively tokenize lines during the first and second
pass of the parser.

Xyce/backlog/project-backlog#480