Remove spice-incompatible breakdown code
Back in commit Xyce/code/uur-proprietary/Xyce@2a543e10 in 2007, a Xyce
developer noticed that the SPICE diode code's algorithm for
reconciling inconsistent user-specified values of BV and
IBV (reverse breakdown knee voltage and reverse breakdown knee
current) could fail to converge, was ignoring the emission
coefficient, and could produce "garbage" values of tBrkdwnV.  He then
wrote a better algorithm and put it into the Xyce diode model.  He
left the original one inside an ifdef that is never enabled and was no
longer tested.

This algorithm was in fact better, but putting it into the legacy
SPICE diode was a violation of the Xyce developer guide, which states
that when implementing a legacy SPICE model, it must match SPICE
exactly, even when it's not the best possible implementation.  We used
to refer to this as "bug compatible".

In fact, the developer guide says:

>What this means is that a designer needs to be confident
>that they can use the same model, and model parameters in code A
>that they used in code B, and still get the same answer.  Switching codes
>should never result in significant changes in the answer.  Minor differences
>due to platform differences, error tolerances, etc., can be acceptable, but
>ultimately, a trusted result from code A should match a result from
>code B.  This means that we need to make the models precisely the same,
>to the extent that this is possible.

Recently, we have been asked to bring in some recent, industry
standard modifications to the original SPICE diode that maintain
backward compatibility but also add additional or improved features.
Doing this in the presence of explicitly spice-incompatible changes is
problematic.  So we are at long last removing this improved breakdown
code and restoring the original method from SPICE.

Making this change broke two Xyce regression tests, both of them using
incompatible BV and IBV parameters.  Those tests, however, were
designed to test Xyce output formatting of HB results and were NOT
meant to be tests of the diode performance.  So the solution there is
to fix their gold standards.

Xyce/backlog/project-backlog#474