Remove spice-incompatible IRF parameter from diode
IRF was a fudge factor added to the Xyce diode by a developer who
intended it to be used by Sandia model card extraction efforts to fine
tune reverse current expressions in Zeners.

As such, it represented a completely unvalidated extension to the Xyce
diode, but in its initial implementation was harmless unless used ---
the default value of IRF=1.0 meant that the diode remained strictly
equivalent to the SPICE3F5 diode it was always intended to be.

A month after implementing IRF, an other completely unvalidated
temperature correction factor was added so that even when IRF was at
its default value, the reverse current expression became incompatible
with SPICE.  At the same time, an incorrect explanation of how it was
applied was added to the Xyce reference guide.

Several years later, a new developer on the Xyce team saw the
discrepancy between the explanation in the reference guide and the
code itself, and rather than fixing the documentation, he "fixed" the
code to match.  This introduced a non-physical, temperature-dependent
discontinuity in the diode current even when IRF was not specified.

The failure of Xyce to fall back to strict SPICE3F5 compliance when
IRF was not specified was fixed in Xyce 6.11 in 2019, but the code was
left in place with a deprecation warning both in the code (if IRF was
specified) and the reference guide.

This commit completely expunges this deprecated "enhancement" from the
diode code.

Xyce/backlog/project-backlog#471