Adds capability to expedite parsing of single-device subcircuits
The second pass of the parser only needs to parse the device lines
and file/circuit hierarchy.  Anything else has already been stored
from the first pass.  A recent change to the parser allowed for
calls to getLine that filter for certain character strings to reduce
the amount of unnecessary tokenizing.  While that illustrated a
performance improvement for the SkyWater PDK, it has also been noted
that modern PDKs are putting single devices in subcircuits together
with models to be used for binning.

This commit leverages that scenario by identifying simple
single-device subcircuits that have no file hierarchy (.INC) or
subcircuits or mutual inductors.  In that case, the parser can
change context to the subcircuit, parse the device line for the
single device, and exit the subcircuit context.  Any excess lines
are not tokenized, as they don't need to be.

This enables the stress test from ISSUE_455 to parse 20x faster and
Xyce does less work when it doesn't need to.

Xyce/backlog/project-backlog#480