Remove lambert-w code from diode
Lambert-W was an experimental technique inserted into the Xyce diode
and BJT as a research project in 2002.  It was interesting at the time
and was explored as an alternative to SPICE-style voltage limiting.

This code has been unused and untested for nearly 20 years.  It is of
only historical interest.

I am removing it in order to clean up the diode so that it will be
easier to read by new developers, easier to maintain, and easier to
update with new, modern SPICE diode effects that have moved beyond
SPICE3F5.

In addition to removing the Lambert-W code, I have also made a few
changes to a handful of conditionals involving "model_.RS" to silence
clang compiler warnings, and to stop using a double as if it were a
boolean (a habit from the very early days of Xyce that was later
abandoned as bad practice).

Xyce/backlog/project-backlog#472