Remove Lambert-W code from BJT
Lambert-W was an experimental effort intended to explore ways to avoid
using SPICE-style voltage limiting in diodes and BJTs.  The reason for
wanting to do that was that this type of limiting is incompatible with
some advanced alternative nonlinear solver methods we wanted to try.

The Lambert-W effort pretty much started and ended in 2002.  It has
been unused and untested since then, and basically just complicates
maintenance of the devices.

It is being removed as part of a code clean-up effort.

Xyce/backlog/project-backlog#472