Changes global to local index when obtaining node name
The local index needs to be used to index into the node name vector
on the processor that owns that node.  The global node was being used
and that was causing a seg fault when the global node was larger than
the length of the local node name vector.

Xyce/backlog/bugs#19