Fix non-standard attribute formatting properly this time
The Verilog-A standard actually says that attributes (parameter or
module) should be comma separated.

ADMS, predating the standard, always has allowed space-separated
attributes.

All of our Verilog-A toy examples that used attributes actually use
space-separated, which is just wrong.  Team members have emulated these
examples to create what they hoped would be portable modeules and
tried them in other simulators,  only to find those simulators puking
on the attributes.

Now the examples are fixed.  Copies of these toys are present in the
VERILOGPLUGIN test case in Xyce_Regression, and those are also getting
fixed.

In a prior commit I had tried to do this and then reverted it claiming
that "adms does not support comma separated attributes."  That was
incorrect.

In fact what was wrong was that years ago I picked up a habit (almost
certainly from some ADMS example I saw once) of defining a macro:
  `define attr(txt) (* txt *)

and then using `attr(attribute1="foo" attribute2="bar") when I
needed to use attributes.  That works just fine in ADMS because ADMS
supports space-separated attributes that are completely non-standard.

Now that we know that the standard requires *comma* separated
attributes, we tried to fix the examples by adding commas.  But that
is the wrong thing to do when using that macro, because it means that
only the first attribute is used, because the macro only takes one
argument and commmas are argument separators.

The thing to do is to stop using this dumb macro.  "(* *)" is no
harder to type than "`attr( )".

Xyce/backlog/project-backlog#486