Remove obsolete issue 1037 from known defects tbl
This should have been removed a really long time ago, that is my
mistake.  The known defect for BUG-1037-SON claimed that "the use
of non-constant values in .PARAM statements may lead to unexpected
results."

It also claimed that this netlist line .PARAM PA = {TEMP} is
forbidden in Xyce since the special variable TEMP is not constant.
It also claimed that the parser would not produce an error if the user
did this.

However, none of this is true at this point.   Bug 1037 was fixed and
closed in 2019, and the resolution of that fix was to have the parser
produce a fatal error for this use case.

Then much later, when the expression library was replaced, Xyce could
handle using special variables in .params.  That was the subject of
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/157, which was
closed about a year ago.

This commit also removes the placeholder entry from the fixed defects
table.