Changes the computation of the startBlock and endBlock
The optimization used by the Windows compilers modified the computation
of the endBlock_ variable such that the answer was incorrect.  This
computation is simple in serial execution and doesn't require such
modification.  Only in parallel execution does this variable need to be
modified.  Furthermore, it really is fine to use simple integer division
as opposed to the more complicated std::floor computation that I opted for.

These two changes were made to the code to address the Windows failure
and make the code more robust.

Xyce/backlog/bugs#16