Add a std::flush to report_handler
On Windows, we would see Certification_Tests/ISSUE_455/issue455 fail
because it was looking for a specific message in stderr but not finding
it. When running this by hand, we would see the error message just fine.
The perl scripts were reditecting stdout and stderr to files. Stderr,
where the error was supposed to appear, turned up empty. This led
us to look at flushing output after printing. Upon inspection,
report_handler in src/CircuitPKG/N_CIR_Xyce.C makes no assumption that
there is a std::flush or a std::endl (which implies flush) in any
messages it prints. This appears to be the cause of the problem we're
seeing on Windows.