Fix bug in the Philips mobility model
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/23

The Philips mobility model had a bug in it that was discovered this
week when it caused some of the weekly TCAD tests to fail.

The model contained a while loop that always (or at least frequently)
produced NaNs.  On most platforms, despite producing NaNs, the while
loop condition would be satisfied, and the function would exit the while
loop and continue on.  The NaNs weren't used, and thus it was a
low-consequence mistake.

However, last week these tests started failing on Windows.  Using the
newest build procedure the optimized Xyce builds were not treating NaNs
in the same way, and thus the while loop condition was *not* treated as
being met. So, the function never escaped the while loop and the
simulation never finished (or got past the initial setup).

This commit fixes the issue.  The function of the while loop was to find
the value for P that produces a minimum value for the function G(P).
P is a screening parameter and G(P) is an analytical function used by
the model to represent minority impurity scattering.  P is dependent on
carrier density, so it is updated at every Newton step.

If P goes negative, then G(P) is NaN.  If P is very small, G(P) gets huge
values which is also numerically problematic.  If you plot G(P), with P
as the x-axis, the function looks like of like an L-shape, with a
minima near 0.3 when Temp=300K.

To avoid the problems near P=0, the model design doesn't allow P
to be smaller than the value that produces this G(P) minima.
However, this minima value isn't a single fixed value, so it
has to be computed.  I've browsed a few open-source TCAD codes, and it
seems they all find this minima in different ways.  It doesn't really
matter how the minima is found, however.

In the old Xyce implementation of this function, the intent was to use
Newton's method to find the minima of G(P).  However, the Newton loop
was set up incorrectly.  It should have been attempting to solve for
G'(P) = 0, where G'(P) is the derivative w.r.t. P.    Doing so would
mean that the update would be given by P -= G'(P)/G''(P), requiring both
the first and second derivative.

Instead it was using G(P) as the residual function, and using
P -= G(P)/G'(P) as the update, which didn't make any sense.  I can't
imagine it ever found a correct solution this way, and the Newton
iterations routinely produced P<0, which caused NaNs.

With this commit I've done the following:

1. re-arranged the Philips mobility model code so it is better organized
   and it is easier to see how it maps to the original paper by Klassen.
2. Used hardwired numbers for Pnmin and Ppmin, when T = 300K.  That way
   the Newton loop is only invoked when T != 300K, which is more
   efficient.
3. Completely rewrote the Newton loop to solve G'(P) = 0.  I am using
   Sacado to get the first and second derivatives of G(P).  Also, the
   loop is now a for-loop rather than a while loop.  So, while it
   will exit early if converged, the max total number of iterations
   has a hard limit.
4. Added an if-statement inside the Newton loop to ensure that P is
   always > 0.  That way, even if the loop stalls it won't produce NaN.
5. Fixed the calling function in the 1D TCAD device (in calcMobilities)
   so that the device temperature is properly passed into the philips
   mobility model.  Before, it unconditionally passed in the default
   value of 300.15K, no matter what had been set in the netlist.  As
   a result, the "while" loop was completely unnecessary.  The model
   already had pre-computed values for Pnmin and Ppmin for 300.15K.
   The while loop only mattered for other temperatures.