Restore orig value of Philips model constant, r5
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/23

In my previous commit on this issue, I fixed various problems in the
Philips mobility model.  One change I forgot to mention was that I
changed the value of a constant, r5, from 0.8552 to 0.01552.

The constant, r5, is used to compute F(P), which is a scattering term in
the Philips model.  When I worked on this model yesterday, I noticed
that the original Klaassen papers used r5=0.01552, while the code I was
editing was using a different value, r5=0.8552.

In at least one open-source code, the 0.01552 value was used so I
decided this was probably the correct one and made the change.  However,
the result was that two TCAD tests started getting compare failures.
These were two of the same tests impacted by the NaN bug that I fixed.
(Not that many of the TCAD tests use this mobility model).

I did a little more digging this morning, and checked the user guide of
every commercial TCAD code I could find.  It looks like all of them use
r5=0.8552, the value Xyce was using before my fix.  Restoring this
constant to that value fixed the comparison failures as well.  So,
although I don't know the pedigree of this change, it is apparently the
correct number to use.  Presumably someone published an update or errata
to the Klaassen 1992 papers with this newer value.