Correct level 2 diode conductances
Since the day it was added to the Xyce diode, the level 2 diode has
incorrectly computed some terms involved in the diode conductances
when IKF and ISATR are set.

These have absolutely no consequence to any test in the test suite,
nor in one of the exceedingly long and slow test cases that actually
uses the level 2 diode.  It must be the case that this error doesn't
really have much impact on the computation.  It doesn't even seem to
have much impact on convergence.

Because they're wrong, I'm fixing them.  It is appalling that they
should have been so wrong for so long.

Xyce/backlog/project-backlog#473