Remove most level=2 conditionals
Almost all of the "level=2" "pspice-compatible" diode stuff reduces to
the level 1 if the extra parameters are left at their
defaults (usually zero).

The level=2 diode code was very primitive and is hardly
"pspice-compatible" --- it recognizes a lot of paramters it doesn't
use, for example.  And the only true incompatibility with the level=1
diode amounts to two lines.

I have removed almost all of the level conditionals except for those
two lines that need to be conditionalized.

Making this change has NO impact on any testing, because any tests
that used the level=1 diode didn't have the level=2 parameters.

Xyce/backlog/project-backlog#473