Remove a bunch of "#if 0"
This is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/281

This commit removes and/or rewrites numerous debug-related ifdefs in the
"new expression" class.

For most cases, the affected code is stuff that would still be useful
when debugging expression issues.  One example is evaluation of the value
and the derivatives of an expression.  This can be accomplished in more
than one way.  Either via the "dx2" function, or via the "val()"
followed by repeated N "dx" function calls, where N=number derivatives

Prior to this change the older way of doing things (val + dx*N)
was behind the #if 0, while the newer, faster (dx2) was uncommented
and used by default.  But occasionally, dx2 will turn out to have a
problem, and so it is convenient to check it vs. the old way.

So, the older code remains, but now instead of using "#if 0" to
comment it out, I have put it inside an "if (false)" conditional.

A good compiler optimization should make any "if(false)" code go
away just like it would w/the preprocessor.  But, with "if (false")
it is at least compile-tested.

There are a number of debug-outputs that I've modified in the same way.
I've also removed some dead code.