Change version on develop branch to 7.7
For the last couple of releases, rather than have develop say it's
some release far in the future (like, release 8), I've been having it
say it's the next release.

That means that when branching for release, we no longer have to
change the release branch's version string --- it's already correct.
Instead, we just update develop to point to one more in the future.

Xyce/backlog/project-backlog#404