Update reference and user guide citations
The UG references the current RG, and vice versa.  It is necessary to
have a bibliography item for each, and for the references to point to
the right one.

Note that I have NOT added a new bibliography entry for XDM 2.6 or
updated references to it.  That document does not yet exist, and the
XDM manual in Xyce/documents/Xyce_Docs> has not been updated since the
2.5 release.

Xyce/backlog/project-backlog#404