Remove "known defects" table
The "known defects" table in the release notes has been around for as
long as we've had release notes, and was originally intended as a list
of bugs we'd found but had not yet had time to fix.

It is woefully out of date, and was never comprehensive.  It has
become a maintenance headache and by general consensus it is being
removed.

Xyce/backlog/project-backlog#404