Have CMake set a default path for lib xycecinterface
The python module xyce_interface.py which accesses
libxyceceinterface takes a path to the library file in
its constructor.  However, Cmake can provide a useful
default value for this path when Xyce is installed.
So, now a working default path is used to load
libxycecinterface or it can be over-ridden if needed.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/477