Add stdbool.h include
bool is not a standard C type, and including math.h is not enough to
get it right on all systems.  It seems to have been OK for RHEL7, but
it's not universal.

I discovered this by trying to run the MIXED_SIGNAL tests on my BSD
system.

According to the standard, in order to use bool in C you must include
stdbool.h.  Apparently this is getting included somewhere through
math.h on RHEL7, but you can't count on that.