Tweak ADMS analysis tool handling of attributes
ADMS comes with a little "back end" that does nothing more than emit
a Verilog-A file.  That is, ADMS reads in a Verilog-A file and creates
its data tree, and this back end processes the data tree and writes it
back out in Verilog-A, with some annotations in comments.

Years ago I created a "xyceVLAtoVLA.xml" script that augmented the
comments output by the one provided with ADMS, because I was looking
into doing some optimization or other and thought it would help me
wrap my brain around the problem.  It didn't.

However today I was thinking about this script and it occurred to me
that it outputs module, parameter, and variable attributes
incorrectly:  it puts them *after* thing they apply to, and it makes
them space separated instead of comma separated.  This flaw was
inherited from the version provided with ADMS.

This commit fixes that flaw.  Attributes now are placed *before* the
item in question and are comma separated, which makes the output of
this script standard compliant (even if the input wasn't!).