Allow the subcircuit "M" parameter to be mutable
"Remove parser error trap for using .global_param on the implicit
subcircuit multiplier parameter (M)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/494

This is a very small commit, which changes a call from:

parameter.getImmutableValue<double>();

to :

double value;
if (parameter.hasExpressionValue())
{
  value = parameter.getMutableValue<double>();
}
else
{ // if not an expression, do things the old-fashioned way
  value = parameter.getImmutableValue<double>();
}

The effect of this is that if the netlist sets a subcircuit "M"
multiplier parameter using a .GLOBAL_PARAM, previously it resulted in a
fatal error and now it does not.

However, this does NOT make the "M" parameter truly mutable.  If the
netlist has a command which will change the value of "M" (via .STEP,
.etc), then it will not work correctly.  The initial value will be used
at every step.  Fixing this will be the subject of another isssue.