Fix for variable subcircuit multipliers
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/492
"Subcircuit multiliplier parameters (M) need to be handled as an AST
(enabling .STEP to be applied to subcircuit M)"

A little while ago, I implemented implicit subcircuit multipliers (the
"M" parmaeter on "X" lines) in Xyce.  However, in doing so, I did it in
an inflexible way.  Whatever value they had at parse time was fixed for
the rest of the simulation.  I think I did this because I was in a
hurry, and then forgot it was like that.

More recently, I attempted to perforrm a .STEP sweep of the subcircuit
multiplier and it didn't work.  The first problem was that there was an
error trap which complained if the multiplier was being set by a
"mutable" parameter.  I removed that error trap in a previous commit.

The next problem was that the multiplier for each subcircuit instance
was being stored in the CircuitContext class as a double precision
number, rather than as part of an abstract syntax tree.  This is fixed
by this commit.

Now, all the multipliers are stored as parameter objects, and their
respective syntax trees are attached to each other as needed to support
sweeping multipliers via commands such as .STEP.  I believe that this
issue is now fixed.