Update error checking in tracking data
For several releases of the Windows version of Xyce, the hostname
field was not populated in the audit/tracking messages that Xyce
posts to a remote server. I was not able to reproduce this problem
and find the root cause, but this commit introduces additional error
checking that might prevent the problem. This change also provides a
means of including error messages in the tracking data sent to the
remote server. If there is still a problem pulling the hostname in the
Windows version of Xyce, these error messages will hopefully help
pinpoint the problem.

Since the tracking data is not typically tested in the regression
suite, this commit also includes a few new unit tests for the
functions used to populate the hostname field in the audit
information.

Xyce/backlog/bugs#18