Stop unconditionally linking to suitesparseconfig
A problem that cropped up in Release 7.6 testing was introduced by
commit d95a894b.  In that commit, I had made configure test
unconditionally for the presence of "libsuitesparseconfig" because in
some rare use cases libamd might require it to be linked
explicitly (the rare use case is that a user has compiled suitesparse
statically from source without relying on a system
package, and that static library would require that
libsuitsparseconfig be statically and explicitly linked --- this is
NOT necessary if libamd is obtained from a properly linked shared
library, because the shared library would pull in libsuitesparseconfig
dynamically).

This hack intended only for one specific (rare) use case caused
problems in our release binaries on RHEL7.  Those are, in fact, built
with a static-linked, VERY OLD version of libamd.a that does NOT
require suitesparseconfig.  But our RHEL7 systems have suitesparse
installed from system packages *in addition* to our static linked one.

Thus, our RHEL7 builds were unnecessarily pulling in libsuitesparse
just because it was found, even though the libamd.a actually being
linked didn't even use it.

That forced all of our RHEL7 RPMs to pull in suitesparse as an
automatic dependency upon installation.

Worse, because our CEE central installs do NOT use RPM and are
installed by hand, they did not automatically cause suitesparse
libraries to be installed, and thus our binaries didn't work on any of
the cee-compute nodes at first (a situation corrected by asking CEE
staff to install it).

This commit refines that kludge.  INstead of blindly searching for
libsuitesparseconfig unconditionally, it now only does so if linking
to libamd fails first.  If that works, we never bother probing for
it.  If it fails, then we look for suitesparseconfig and try libamd
again.  if *that* works, great, we use both, and if that fails then we
barph saying that we could not get libamd linked.

I have tested this on our RHEL7 machines and it no longer pulls in an
unnecessary libsuitesparseconfig.  On my BSD machine, where libamd.so
is already properly linked up so it pulls in libsuitesparseconfig
dynamically, the first probe of libamd works fine and so we don't
probe for libsuitesparseconfig at all.

So now, the probe for libsuitesparseconfig won't ever happen unless it
turns out that someone is using a very new libamd.a *and* has no
shared library version of it.

I know that autoconf is slated for destruction upon full roll-out of
CMake, but I couldn't leave it in this state.  I'd have fixed it
earlier if it had been detected prior to building our 7.5 QA builds.
Once we got to that stage, though, it was too late to fix it.