Updated method signature on test function
I had created a pseudo-mock function to substitute a system call for
use in a unit test. On the system I developed on, the function
prototype I copied included a throw() attribute. This form is
apparently somewhat unique, and the unit test code failed to build on
Mac system. This commit removes the attribute.