Allow ceil and floor to be used in analog functions
An oversight in the handling of analog function derivatives led to the
functions "ceil" and "floor" being considered invalid in analog
functions.  There is nothing wrong with using them, and the only
problem was that the derivative code was not set up to treat them
specially (and they are special, so need that).

This commit corrects the oversight.
Xyce/backlog/bugs#35