Allow Xyce/ADMS to ignore callfunctions in analog functions
$strobe, $bound_step, $write, etc. are known to ADMS as
"callfunctions".  They are statements that look like function calls (not
assignments with function calls on the RHS) and do stuff like throw
output, specify a maximum time step size, etc.

These are illegal in Xyce analog functions, because we generate the
functions as being outside of the device class and things like
UserInfo (that we output when $strobe is used normally) need an
instance pointer to work, and $bound_step needs access to instance
data members.

Until this commit, Xyce/ADMS would abort with an error about a missing
template.  Simply adding the template wasn't enough, as then invalid
code got generated.

Now, the call functions are ignored, and Xyce/ADMS outputs a warning
that it is doing so.

Xyce/backlog/bugs#36