Update agauss/gauss so 3rd argument is optional
This is a minor enhancement.  The expression library supports gauss and
agauss random operators, but they were written to require 3 arguments.
For the sake of compatibility, the 3rd argument needs to be optional,
with a default value of 1.0.  This commit fixes this minor issue.

This commit also creates a handful of unit tests demonstrating the
feature.

Also, in this commit is the beginnings of code to support an optional
4th argument to gauss/agauss, and optional 3rd argument to the other
random operators such as unif and aunif.  This new optional argument is
the multiplier.  This commit supports it syntatically, but doesn't yet
do anything with it.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/79
"Make "n" (AKA num_sigma) parameter optional for AGAUSS and GAUSS"