Improves robustness of the KSparse solver to forward and transpose linear solves
Several issues have been found in KSparse that are impacting its robustness
and ability to perform transpose solves:

1. The manner in which the matrix is copied from Epetra into the KSparse data
format is not compatible with the transform interface, which is used for
parallel distribution, singleton filtering, etc.  Thus, linear solver
transformations have been disabled in KSparse.  For a sparse-direct solver
this is not that big of a problem.  It is more of a problem for iterative
solvers.

2.  The Epetra-KSparse interface only performs reordering and matrix scaling
for the first linear solve and then reuses it throughout the rest of the
lifetime of the solver object.  That has never been robust enough for circuit
simulation.  Reordering and scaling is now being performed for every solve.

3.  Issues with transpose solves has been tracked down to an issue with the
use of the reordering map after the solve has been performed.  It was using
the row map when it should be using the column map.  That has been fixed.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/288