Make TRS an alias for TRS1
Other simulators (HSPICE, ngspice) allow the diode model parameter
"TRS" to be an alias for TRS1.  We did not, and it has bitten users

While it is unrelated to the issue I'm working on, I'm adding that
now.

I am also enabling analytic sensitivities on the TRS* parameters,
because they were in fact already properly being handled in all the
sensitivity code (the code was checking if the parameter requested was
one of the TRS variables, and if so, it was using it as the
independent variable for differentiation).